restructures Nav component
file structure for Nav component updated. Things concerned with the component, like styles, are collocated
