+++ src/components/PageContent.js
@@ -7,8 +7,8 @@ export default function PageContent(props) {
 
     const styles = {
         backgroundColor: isDarkMode? "black" : "white", 
-        height: "100vh", 
-        width: "100vw",
+        height: "100%", 
+        width: "100%",
     }
     return <div style={styles}>{props.children}</div>
 }
